🌿 First steps to convert from Blocking to Reactive - We keep the pagination for now

+++ stamps-and-coins-blocking-service/src/main/kotlin/org/jesperancinha/enterprise/staco/jpa/service/StaCoService.kt
@@ -59,8 +59,8 @@ class StaCoService(
         return ResponseDto(
             staCoDtos = contentEntities.map { it.toDto },
             currentPage = pageEntities,
-            totalRecords = searchEntities.numberOfElements,
-            totalPages = searchEntities.totalPages
+            totalRecords = searchEntities.numberOfElements.toLong(),
+            totalPages = searchEntities.totalPages.toLong()
         )
     }
 }
