+++ stamps-and-coins-common/src/main/kotlin/org/jesperancinha/enterprise/staco/common/dto/ResponseDto.kt
@@ -3,6 +3,6 @@ package org.jesperancinha.enterprise.staco.common.dto
 data class ResponseDto(
     val staCoDtos: List<StaCoDto>,
     val currentPage: Int,
-    val totalRecords: Int,
-    val totalPages: Int,
+    val totalRecords: Long,
+    val totalPages: Long,
 )
